requirements.txt
streamlit
password-strength
zxcvbn-python